import cleaning, small refactoring and add doc
